Change Ecosystem to incorporate streams
Ecosystems now store a set of streams that allow for the transport of guppies
between pools. To adjust for crowding, the weakest guppies will now be
transported between pools through streams instead of dying. If there are
no further pools to travel to through streams, then the guppies will die
to overcrowding.

~TODO~
Ecosystem's adjustForCrowding method must use the Stream class to
transport the guppies. The guppies dying when there are no available pools
to travel to has already been implemented.

Pool's adjustForCrowding method must return the set of the weakest guppies
that need to be dealt with to fix overcrowding.