Handle overcrowding by transporting guppies
Ecosystem transports guppies through the randomly selected valid stream
to the next pool. If there are no further pools, the guppies die
instead. Adding a pool to the ecosystem also creates a stream from the
last pool to the new pool. A similar functionality sets up streams
between pools when the ecosystem's pools are explicitly set. This
functionality should be removed once streams are set up to be
non-linear.

Pool now handles overcrowding by removing the weakest guppies from the
pool and returning the set of weakest guppies.

Streams can now transport guppies to their destination pool. A random
chance has been implemented to determine whether the guppies dies in
transport.

~TODO~
Guppies that die in transport should be tracked and this count should
be returned.