Track number of guppies that die in transport
Stream's transportGuppies method now returns the number of guppies that
die in transport.

Ecosystem's adjustForOvercrowding method adds the number that die in
transport to the count that it returns (the number of guppies that die
due to overcrowding).

~TODO~
Guppies that die in transport die as a indirect result of overcrowding,
but their explicit cause of death is dying in transport. Should this be
handled otherwise? Add an issue to keep track of this.