Refactor for performance
The manual sorting method in Pool (sortGuppiesByHealthCoefficient) was
causing the application to hang as it went through its long process. This
changes the adjustForCrowding method in Pool to use the sorting method
from Collections instead. This has greatly improved performance.

Guppy was modified to implement Comparable in order to allow this sorting
method to be used.

This commit also fixes a concurrentModificationException occurring from
the spawn method in Pool.