Restructure for JavaFX GUI
The program has been restructured to implement the JavaFX GUI. The point
of entry into the program is now SimulationApplication.java.

Currently, the program will load the GUI with a message to load a file. A
file can be selected using the 'load' button. This file will then be
loaded into the simulation and a visual representation will be displayed
in the GUI. This is the extent of the GUI functionality at this point.

~Todo~
- Implement stepping simulation (simulate one week/restore data from
  history if it exists)
- Implement stepping simulation backwards (return to previous week in
  history