Implement stepping one week forward or backward
The simulation can now be stepped one week forwards or backwards. The
simulation stores a history of the state at the beginning of each week.
This is used to return to previous states of the simulation.

To implement stepping, copy methods needed to be created to create a copy
of the ecosystem to store in history.
The following classes now contain a copy method:
- Ecosystem
- Pool
- Stream
- Guppy

The animation pane will now only draw up to a maximum of 1000 guppies per
pool. This is to reduce visual clutter and to reduce hang time.

The GUI has also been configured further to disable buttons when they do not
have a use.