Remove redundant copy
For simulating new weeks, a copy is made of the previous week's ecosystem
in order to simulate. After simulating, a copy of the first copy was then
made to store in history. This second copy was not needed. The original
copy that undergoes simulation is now stored in history instead.