Implement play/pause functionality
The simulation can now be automatically run using the play/pause button.
By default, the simulation will advance one week every one second. This
can be changed using the speed slider.

Beware running the simulation past week 15; it starts to take a few
seconds to simulate the week. If you are on a faster speed, the simulation
may freeze while it tries to catch up with the scheduled simulations.

~TODO~
- Add status pane to the left side of the GUI to display current
  information
- Add data pane to the right side of the GUI to display
  statistics/graphs/etc
- Hopefully improve simulation algorithm so that it does not take
  noticeable amounts of time past week 15
  - maybe a certain number of weeks can be simulated in advance