Refactor to use MVC structure
The application now uses a Model-View-Controller structure to
separate application and GUI logic.

The Simulation object is the main model.
The GUI object is the main view.
The SimulationController object is the main controller.

The controls in the GUI have been modified to be handled
by explicit handlers. This allows for further separation of
specialized logic.