Add update functionality to graph
An update button allows the user to update the graph to the current
simulation state. Now, you can open the graph, continue the simulation,
and then update the graph instead of having to open a new graph window.

An auto-update option has also been added; however, this results in a
ConcurrentModificationException within a few weeks of running, stopping
the auto-update. The auto-update option must be toggled off, and then on
again to continue auto-updating.

The 'settings' menu in the graph window has been refactored to be an
'options' menu.