Graph data is updated by the Simulation
Instead of the graph handler possibly requesting a copy of the simulation
history while it is being modified, the simulation will instead update the
handler with a new copy of the history when it is finished processing each
interval (currently 1 week). This fixes the
ConcurrentModificationException that occurred with the previous
implementation.