Add graph saving functionality
The graph can now be exported as a PNG image using the save button in the
graph window.

A bug preventing the graph to grow horizontally has been fixed by setting
the Hgrow property when the graph is added instead of separately. This
issue occurred because the graph had not been added when the Hgrow
property was being set due to the graph being added on the JavaFX thread.

The option pane on the graph window now has a minimum width, preventing
the controls from being collapsed to unreadable sizes when resizing the
window.