Add a third-dimension to the animation
This replaces the 2D animation with a 3D representation.

The 3D view can be rotated by left button dragging, panned by right button
dragging, and zoomed in and out by scrolling up and down.

The view can be reset to the default by pressing 'R'.

Currently, the 3D view consists of 200 spheres moving around each 'pool'.
This will be changed to represent guppies in the near future.

The DualLayerPane implementation has been replaced with explicit calls to
set the background colour.

The root node of the application has been changed to an anchor pane to
allow for proper resizing of inner nodes when the application window is
resized.

Each option control has been given a minimum width to prevent shrinking
past the minimum, making the control text unreadable.