Fix intermittent transparency loss
The 3D boxes (pools) were originally being loaded from right to left. This
was not noticed until a dataset with > 3 pools was used. This was fixed by
inverting the direction along which pools were laid out; however, this
meant that pools were added in the opposite order than required for
transparency to work properly.

This was fixed by reversing when the order needed to be reversed. This led
to the issue that newly created nodes were not reversed when they needed
to be. To deal with this, newly created ecosystem representations will
have their nodes reversed if the boolean flag is true within the SubScene
object.