Make controller store a copy of simulation history
The controller now stores the copy of the simulation history. Instead of
regenerating a copy of the entire history every time a new week is
simulated, a copy of the newest simulation state is added to the
controller's stored copy of the history. This change resulted in massive
performance gains.